Guide UTA setup during onboarding
